modal and tooltip added
