+++ src/components/Common/Constants.tsx
@@ -1,0 +1,82 @@
+import { Button, Input, Modal } from "antd";
+import React, { useState } from "react";
+import Loader from "./Loader";
+import PrimaryButton from "./PrimaryButton";
+import SecondaryButton from "./SecondaryButton";
+import { LoadingOutlined } from "@ant-design/icons";
+import { StyledCard } from "./Styles";
+
+export default function Constants() {
+  // const [visible, setVisible] = useState(false);
+  // const  handleOk = () => {
+  //   setVisible(false);
+  // };
+  // const handleCancel = () => {
+  //   setVisible(false);
+  // };
+  // const showModal = () => {
+  //   setVisible(true);
+  // };
+
+  // const antIcon = <LoadingOutlined style={{ fontSize: 24 }} spin />;
+
+  // const data = [
+  //   {
+  //     name: "Button",
+  //     component: <Button>Submit</Button>,
+  //     key: "1",
+  //   },
+  //   {
+  //     name: "Input",
+  //     component: <Input placeholder={"enter your name"}></Input>,
+  //     key: "2",
+  //   },
+  //   {
+  //     name: "Primary Button",
+  //     component: <PrimaryButton>Sign Up</PrimaryButton>,
+  //     key: "3",
+  //   },
+  //   {
+  //     name: "Secondary button",
+  //     component: <SecondaryButton>Cancel</SecondaryButton>,
+  //     key: "4",
+  //   },
+  //   {
+  //     name: "Card",
+  //     component: (
+  //       <StyledCard bordered={false} hoverable={true}>
+  //         {" "}
+  //         This is Card Component{" "}
+  //       </StyledCard>
+  //     ),
+  //     key: "5",
+  //   },
+  //   {
+  //     name: "Loader",
+  //     component: <Loader icon={antIcon} spinning={false}></Loader>,
+  //     key: "6",
+  //   },
+  //   {
+  //     name: "Modal",
+  //     component: <Modal onOk={handleOk} onCancel={handleCancel}></Modal>,
+  //     key: "7",
+  //   },{
+  //       name:""
+  //   }
+  // ];
+
+  // const columns = [
+  //   {
+  //     title: "Name",
+  //     dataIndex: "name",
+  //     key: "key",
+  //   },
+  //   {
+  //     title: "Component",
+  //     dataIndex: "component",
+  //     key: "key",
+  //   },
+  // ];
+
+  return;
+ }
