+++ src/components/Common/SecondaryButton.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import { StyledSecondaryBtn } from './Styles';
-function SecondaryBtn(props:any) {
+import {StyledSecondaryBtn} from './Styles';
+function SecondaryButton(props:any) {
     const {children}=props;
   return (
      <StyledSecondaryBtn>{children}</StyledSecondaryBtn>
   )
 }
 
-export default SecondaryBtn
+export default SecondaryButton
