--- src/components/ProButton.css
@@ -1,32 +1,0 @@
-/* .UIButton{
-
-    color: aliceblue;
-    background-color: cadetblue;
-}
-
-.UIButton2{
-
-    color: rgb(14, 230, 21);
-    background-color: pink;
-}
-.UIButton3:hover{
-   background-color: blueviolet;
-}
-/* .button-Primary{
-background-color: skyblue;
-}
-.button-Secondary{
-    background-color: lavender;
-
-} */
-/* .btn-Primary{
-    background-color: skyblue;
-    margin-right: 20px;
-border-radius: 3px;
-}
-.btn-Secondary{
-  border-color: skyblue;
-  border-radius: 3px;
-  margin-right: 20px;
-  background-color: aliceblue;
-} */
