+++ src/components/SignIn.css
@@ -1,57 +1,130 @@
-.signInUsername{
-   margin-top: 5px;
-   width: 100%;
-   
-   height: 5vh;
-}
-.signInPass{
-    margin-top: 5px;
-   width: 100%;
-   height: 5vh;
-}
-.signIncard{
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
-  box-sizing: border-box;
-  position: relative;
-  display: flex;
+.signin-div {
+ 
+  background: url(background.png);
+  background-size: cover;
+  background-attachment: fixed;
   justify-content: center;
   align-items: center;
-  min-height: 600px;
-  width: 500px; 
+  display: flex;
+  height: 100vh;
+
 }
 
-.signInImg{
-   height: 100px;
-   width: 100px; 
-   display: block;
-   margin-left: auto;
-   margin-right: auto;
-}
-.h2{
-  text-align: center;  
-  font-weight:800;
-}
-.signInButton{
-   margin-top: 40px;
-   width: 100%;
-   height: 50px;
-   background-color: #227F99;
-color: white;
+.signin-card {
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
+  box-sizing: border-box;
+  display: flex;
+  justify-content: center;
+  max-width: 50vh;
+  border-radius: 0px;
+ 
 }
 
-.signIn{
-   background-color:#227F99;
-   justify-content: center;
-   align-items: center;
-   display: flex;
-   height: 100vh;
-   padding: 24px;
+.signin-card .ant-card-body {
+  width: fit-content !important;
 }
-.envimg{
-   display: flex;
-   justify-content: right;
-   /* text-align: right; */
-   position: absolute;
-   right: 0;
-   top:-1px;
+
+.signin-username {
+  margin-top: 5px;
+  height: 4vh;
+  border: none;
+  margin-bottom: 1px;
+  border-bottom: 1px solid #333;
+  border-radius: 0;
 }
+
+.signin-username:focus {
+  border-bottom: 1px solid dimgray;
+  box-shadow: 0 1px 0 0 dimgray;
+}
+
+.signin-password:focus {
+  border-bottom: 1px solid dimgray;
+  box-shadow: 0 1px 0 0 dimgray;
+}
+
+.signin-password {
+  margin-top: 5px;
+  height: 4vh;
+  border: none;
+  margin-bottom: 1px;
+  border-bottom: 1px solid #333;
+  border-radius: 0;
+}
+
+.signin-button {
+  border: none;
+  margin-top: 5px;
+  width: 100%;
+  height: 35px;
+  background-color: #129489;
+  color: white;
+  margin-bottom: 10px;
+  border-radius: 0;
+}
+
+.signin-image {
+  height: 138px;
+  width: 130px;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.signin-heading {
+  text-align: center;
+  text-align: center;
+  font-weight: bolder;
+  color: #129489;
+  font-size: 32px;
+}
+
+.signin-link {
+  color: #129489;
+  display: block;
+  margin-bottom: 10px;
+  font-size: 15px;
+}
+
+.ant-card .ant-card-body {
+  width: 55vh;
+}
+
+.signin-sub-heading {
+  margin-top: none;
+  text-align: center;
+  font-weight: lighter;
+  font-size: 15px;
+}
+
+.signin-checkbox {
+  margin-left: 10px;
+}
+
+.signin-username:hover {
+  border-bottom-color: #129489;
+}
+
+.signin-password:hover {
+  border-bottom-color: #129489;
+}
+
+@media only screen and (max-width: 1920px) {
+  .signin-card {
+    width: 500px;
+  }
+}
+
+@media only screen and (max-width: 1280px) {
+  .signin-card {
+    width: 400px;
+  }
+}
+
+@media only screen and (max-width: 360px) {
+  .signin-card {
+    width: 340px;
+  }
+}
+
+
